+++ src/TestSuite/CellTestCase.php
@@ -16,47 +16,51 @@ declare(strict_types=1);
 
 namespace MeCms\TestSuite;
 
+use Cake\Http\ServerRequest;
 use MeCms\View\Helper\WidgetHelper;
 use MeCms\View\View\AppView as View;
 
 /**
  * Abstract class for test cells
+ * @property \MeCms\View\Helper\WidgetHelper $Widget
+ * @todo renamed as `WidgetTestCase`
  */
 abstract class CellTestCase extends TestCase
 {
     /**
-     * @var \MeCms\View\Helper\WidgetHelper
-     */
-    protected WidgetHelper $Widget;
-
-    /**
-     * If `true`, a mock instance of the shell will be created
-     * @var bool
-     */
-    protected bool $autoInitializeClass = true;
-
-    /**
-     * Called before every test method
-     * @return void
+     * Get magic method.
+     *
+     * It provides access to the cached properties of the test.
+     * @param string $name Property name
+     * @return mixed
      * @throws \ReflectionException
      */
-    protected function setUp(): void
+    public function __get(string $name)
     {
-        parent::setUp();
+        switch ($name) {
+            //Rewrites the parent's method
+            case 'alias':
+                if (empty($this->_cache['alias'])) {
+                    $this->_cache['alias'] = substr($this->getAlias($this), 0, -7);
+                }
 
-        if (empty($this->Widget) && $this->autoInitializeClass) {
-            $this->Widget = new WidgetHelper(new View());
+                return $this->_cache['alias'];
+            //Rewrites the parent's method
+            case 'Table':
+                if (empty($this->_cache['Table'])) {
+                    $className = $this->getTableClassNameFromAlias($this->alias);
+                    $this->_cache['Table'] = $this->getTable($this->alias, compact('className'));
+                }
+
+                return $this->_cache['Table'];
+            case 'Widget':
+                if (empty($this->_cache['Widget'])) {
+                    $this->_cache['Widget'] = new WidgetHelper(new View((new ServerRequest())->withEnv('REQUEST_URI', '/')));
+                }
+
+                return $this->_cache['Widget'];
         }
 
-        $request = $this->Widget->getView()->getRequest()->withEnv('REQUEST_URI', '/');
-        $this->Widget->getView()->setRequest($request);
-
-        if (empty($this->Table) && $this->autoInitializeClass) {
-            $alias = substr($this->getAlias($this), 0, -7);
-            $className = $this->getTableClassNameFromAlias($alias);
-            if (class_exists($className)) {
-                $this->Table = $this->getTable($alias, compact('className'));
-            }
-        }
+        return parent::__get($name);
     }
 }